handle new build location: default -> Release; fixed typo: DNSSD -> DNS_SD; reformatted list of source files

+++ wscript
@@ -32,26 +32,29 @@ def configure(conf):
 
 
 def post_build(ctx):
-  #print("=== post")
   if not os.path.exists('lib/binding.node'):
+    if os.path.exists('build/Release/binding.node'):
+      os.symlink( '../build/Release/binding.node', 'lib/binding.node')
+    else:
       os.symlink( '../build/default/binding.node', 'lib/binding.node')
 
 def build(bld):
   bld.add_post_fun(post_build)
   obj = bld.new_task_gen('cxx', 'shlib', 'node_addon')
   obj.target = 'binding'
-  obj.uselib = 'DNSSD'
-  obj.includes = '.'
-  obj.source = ['src/binding.cpp',
-                'src/dns_service_browse.cpp',
-                'src/dns_service_enumerate_domains.cpp',
-                'src/dns_service_process_result.cpp',
-                'src/dns_service_ref.cpp',
-                'src/dns_service_ref_deallocate.cpp',
-                'src/dns_service_ref_sock_fd.cpp',
-                'src/dns_service_register.cpp',
-                'src/dns_service_resolve.cpp',
-                'src/mdns_utils.cpp']
+  obj.uselib = 'DNS_SD'
+  obj.includes = ['.']
+  obj.source = [ 'src/binding.cpp'
+               , 'src/dns_service_browse.cpp'
+               , 'src/dns_service_enumerate_domains.cpp'
+               , 'src/dns_service_process_result.cpp'
+               , 'src/dns_service_ref.cpp'
+               , 'src/dns_service_ref_deallocate.cpp'
+               , 'src/dns_service_ref_sock_fd.cpp'
+               , 'src/dns_service_register.cpp'
+               , 'src/dns_service_resolve.cpp'
+               , 'src/mdns_utils.cpp'
+               ]
   if bld.env.HAVE_DNSSERVICEGETADDRINFO:
       obj.source.append('src/dns_service_get_addr_info.cpp')
